Create dbo.item_employee_assignment
the purpose of this table is to store information of employee and the items he/she is accountable for. This table is also the part of the system where monitoring of items are performed. The table includes the item_id (foreign key and the primary key of tblitem table), employee_id (foreign key and the primary key of tblemployee table) and the condition/remarks of the items.